Package classes, removed ADD from Instruction Set
Package entire program into one, also packaged the Task, Todo, Deadline, Event class into its own package/
add and todo used to do the same thing, so add instruction is subsumed under todo instruction.
